move #!who and #!whos to a single implementation that operates on DotNetKernel
